refactor(web): put lib/tools on the standard layout
The directory named its files after what they were about rather than what
they are, so an API handler could be `mcpService.ts` or `openApiService.ts`
and a reader had to open both to find out.

  mcpService.ts     -> svc.ts
  openApiService.ts -> svc.ts (calls) + utils.ts (parsing)
  fetchTools.ts     -> svcSS.ts
  mcpUtils.tsx      -> utils.ts

`openApiService.ts` was two things: four API calls and a pure OpenAPI
definition parser. They go to svc.ts and utils.ts respectively.

`mcpUtils` was a `.tsx` holding no JSX — its only angle bracket is a type
argument — so it becomes `.ts`.
