+++ web/src/app/craft/v1/apps/admin/McpServerPolicyModal.tsx
@@ -4,7 +4,7 @@ import useSWR from "swr";
 import { errorHandlingFetcher } from "@/lib/fetcher";
 import { SWR_KEYS } from "@/lib/swr-keys";
 import { MCPServer, ToolSnapshot } from "@/lib/tools/types";
-import { updateMCPServer } from "@/lib/tools/mcpService";
+import { updateMCPServer } from "@/lib/tools/svc";
 import ActionPolicyEditorModal from "@/app/craft/v1/apps/admin/ActionPolicyEditorModal";
 import type { EndpointPolicy } from "@/app/craft/v1/apps/registry";
