+++ web/src/app/craft/v1/tasks/components/PreApprovalSummary.tsx
@@ -3,7 +3,7 @@
 import { Tag, Text } from "@opal/components";
 import useUserExternalApps from "@/hooks/useUserExternalApps";
 import { useCraftMcpServers } from "@/lib/tools/hooks";
-import { getActionIcon } from "@/lib/tools/mcpUtils";
+import { getActionIcon } from "@/lib/tools/utils";
 import { getAppTypeLogo } from "@/app/craft/v1/apps/registry";
 
 interface PreApprovalSummaryProps {
