+++ web/src/lib/skills/pickerIcons.ts
@@ -1,7 +1,7 @@
 import type { IconFunctionComponent } from "@opal/types";
 import { SvgSparkle } from "@opal/icons";
 import { getAppTypeLogo } from "@/app/craft/v1/apps/registry";
-import { getActionIcon } from "@/lib/tools/mcpUtils";
+import { getActionIcon } from "@/lib/tools/utils";
 import type { PickerEntry } from "@/lib/skills/picker";
 
 /** Icon for a picker entry: the provider logo for apps, the server logo for MCP,
