--- web/src/lib/tools/openApiService.ts
@@ -1,171 +1,0 @@
-import { MethodSpec, ApiResponse, ToolSnapshot } from "@/lib/tools/types";
-
-const SUPPORTED_HTTP_METHODS = new Set([
-  "get",
-  "post",
-  "put",
-  "patch",
-  "delete",
-  "options",
-  "head",
-]);
-
-const isPlainRecord = (value: unknown): value is Record<string, any> =>
-  Boolean(value) && typeof value === "object" && !Array.isArray(value);
-
-export function extractMethodSpecsFromDefinition(
-  definition?: Record<string, any> | null
-): MethodSpec[] {
-  if (!isPlainRecord(definition) || !isPlainRecord(definition.paths)) {
-    return [];
-  }
-
-  const pathEntries = Object.entries(definition.paths as Record<string, any>);
-  const methods: MethodSpec[] = [];
-
-  for (const [path, operations] of pathEntries) {
-    if (!isPlainRecord(operations)) {
-      continue;
-    }
-
-    for (const [methodName, spec] of Object.entries(operations)) {
-      if (!isPlainRecord(spec)) {
-        continue;
-      }
-
-      if (!SUPPORTED_HTTP_METHODS.has(methodName.toLowerCase())) {
-        continue;
-      }
-
-      const name = spec.operationId ?? spec.operationID;
-      const summary = spec.summary ?? spec.description;
-
-      if (!name || !summary) {
-        continue;
-      }
-
-      methods.push({
-        name,
-        summary,
-        path,
-        method: methodName.toUpperCase(),
-        spec,
-        custom_headers: [],
-      });
-    }
-  }
-
-  return methods;
-}
-
-export async function validateToolDefinition(toolData: {
-  definition: Record<string, any>;
-}): Promise<ApiResponse<MethodSpec[]>> {
-  try {
-    const response = await fetch(`/api/admin/tool/custom/validate`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(toolData),
-    });
-
-    if (!response.ok) {
-      const errorDetail = (await response.json()).detail;
-      return { data: null, error: errorDetail };
-    }
-
-    const responseJson = await response.json();
-    return { data: responseJson.methods, error: null };
-  } catch (error) {
-    console.error("Error validating tool:", error);
-    return { data: null, error: "Unexpected error validating tool definition" };
-  }
-}
-
-export async function createCustomTool(toolData: {
-  name: string;
-  description?: string;
-  definition: Record<string, any>;
-  custom_headers: { key: string; value: string }[];
-  passthrough_auth: boolean;
-}): Promise<ApiResponse<ToolSnapshot>> {
-  try {
-    const response = await fetch("/api/admin/tool/custom", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(toolData),
-    });
-
-    if (!response.ok) {
-      const errorDetail = (await response.json()).detail;
-      return { data: null, error: `Failed to create tool: ${errorDetail}` };
-    }
-
-    const tool: ToolSnapshot = await response.json();
-    return { data: tool, error: null };
-  } catch (error) {
-    console.error("Error creating tool:", error);
-    return { data: null, error: "Error creating tool" };
-  }
-}
-
-type ToolUpdatePayload = {
-  name?: string;
-  description?: string;
-  definition?: Record<string, any>;
-  custom_headers?: { key: string; value: string }[] | null;
-  passthrough_auth?: boolean;
-  oauth_config_id?: number | null;
-};
-
-export async function updateCustomTool(
-  toolId: number,
-  toolData: ToolUpdatePayload
-): Promise<ApiResponse<ToolSnapshot>> {
-  try {
-    const response = await fetch(`/api/admin/tool/custom/${toolId}`, {
-      method: "PUT",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(toolData),
-    });
-
-    if (!response.ok) {
-      const errorDetail = (await response.json()).detail;
-      return { data: null, error: `Failed to update tool: ${errorDetail}` };
-    }
-
-    const updatedTool: ToolSnapshot = await response.json();
-    return { data: updatedTool, error: null };
-  } catch (error) {
-    console.error("Error updating tool:", error);
-    return { data: null, error: "Error updating tool" };
-  }
-}
-
-export async function deleteCustomTool(
-  toolId: number
-): Promise<ApiResponse<boolean>> {
-  try {
-    const response = await fetch(`/api/admin/tool/custom/${toolId}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-
-    if (!response.ok) {
-      const errorDetail = (await response.json()).detail;
-      return { data: false, error: `Failed to delete tool: ${errorDetail}` };
-    }
-
-    return { data: true, error: null };
-  } catch (error) {
-    console.error("Error deleting tool:", error);
-    return { data: false, error: "Error deleting tool" };
-  }
-}
