+++ web/src/sections/actions/OpenApiActionCard.tsx
@@ -7,8 +7,8 @@ import ToolsList from "@/sections/actions/ToolsList";
 import { useCreateModal } from "@opal/components";
 import { ToolSnapshot, ActionStatus, MethodSpec } from "@/lib/tools/types";
 import ToolItem from "@/sections/actions/ToolItem";
-import { extractMethodSpecsFromDefinition } from "@/lib/tools/openApiService";
-import { updateToolStatus } from "@/lib/tools/mcpService";
+import { extractMethodSpecsFromDefinition } from "@/lib/tools/utils";
+import { updateToolStatus } from "@/lib/tools/svc";
 import { can } from "@/lib/permissions/resource-actions";
 import { SvgServer, SvgTrash } from "@opal/icons";
 import { ConfirmationModalLayout as Modal } from "@opal/layouts";
