+++ web/src/sections/actions/OpenApiPageContent.tsx
@@ -12,8 +12,8 @@ import AdminListHeader from "@/sections/admin/AdminListHeader";
 import { toast } from "@opal/layouts";
 import OpenApiActionCard from "./OpenApiActionCard";
 import { createOAuthConfig, updateOAuthConfig } from "@/lib/oauth/api";
-import { updateCustomTool, deleteCustomTool } from "@/lib/tools/openApiService";
-import { updateToolStatus } from "@/lib/tools/mcpService";
+import { updateCustomTool, deleteCustomTool } from "@/lib/tools/svc";
+import { updateToolStatus } from "@/lib/tools/svc";
 import DisconnectEntityModal from "./modals/DisconnectEntityModal";
 import ActionCardSkeleton from "./skeleton/ActionCardSkeleton";
 import useOpenApiTools from "@/hooks/useOpenApiTools";
